Fixing IDENTITY-4297 (no support issue as of yet).

+++ components/identity-mgt/org.wso2.carbon.identity.mgt/src/main/java/org/wso2/carbon/identity/mgt/services/UserInformationRecoveryService.java
@@ -70,6 +70,7 @@ import java.util.Map;
  */
 public class UserInformationRecoveryService {
 
+    private static final String USER_ALREADY_EXISTING = "UserAlreadyExisting";
     Log log = LogFactory.getLog(UserInformationRecoveryService.class);
 
     public CaptchaInfoBean getCaptcha() throws IdentityMgtServiceException {
@@ -857,7 +858,7 @@ public class UserInformationRecoveryService {
             vBean = UserIdentityManagementUtil.getCustomErrorMessages(e, userName);
             //Rollback if user exists
             try {
-                if (userStoreManager.isExistingUser(userName)) {
+                if (!e.getMessage().contains(USER_ALREADY_EXISTING) && userStoreManager.isExistingUser(userName)) {
                     userStoreManager.deleteUser(userName);
                 }
             } catch (org.wso2.carbon.user.core.UserStoreException e1) {
